requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-jose==3.4.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0